:recycle: refactor(test): 단위 테스트 폴더를 관리자 메뉴에 따라 분리
